+++ bin/create_alexa_audio_handler.py
@@ -238,13 +238,13 @@ def main(argv):
     try:
         opts, args = getopt.getopt(argv, "hr:", ["root="])
     except getopt.GetoptError:
-        print 'create_alexa_handler.py -r <root project dir>'
-        print 'if -r option not supplied it will look for PWD environment variable'
+        print('create_alexa_handler.py -r <root project dir>')
+        print('if -r option not supplied it will look for PWD environment variable')
         sys.exit(2)
     for opt, arg in opts:
         if opt == '-h':
-            print 'create_alexa_handler.py -r <root project dir>'
-            print 'if -r option not supplied it will look for PWD environment variable'
+            print('create_alexa_handler.py -r <root project dir>')
+            print('if -r option not supplied it will look for PWD environment variable')
             sys.exit()
         elif opt in ("-r", "--root"):
             root_project_dir = arg