+++ bin/create_alexa_test_skills.py
@@ -35,14 +35,14 @@ def main(argv):
     try:
         opts, args = getopt.getopt(argv, "hr:", ["root="])
     except getopt.GetoptError:
-        print 'create_alexa_test_skills.py -r <root project dir>'
-        print 'if -r option not supplied it will look for PWD environment variable'
+        print('create_alexa_test_skills.py -r <root project dir>')
+        print('if -r option not supplied it will look for PWD environment variable')
         sys.exit(2)
     for opt, arg in opts:
         if opt == '-h':
-            print 'create_alexa_test_skills.py -r <root project dir>'
-            print 'if -r option not supplied it will look for PWD environment variable'
-            print '<include files> are relative to root project dir'
+            print('create_alexa_test_skills.py -r <root project dir>')
+            print('if -r option not supplied it will look for PWD environment variable')
+            print('<include files> are relative to root project dir')
             sys.exit()
         elif opt in ("-r", "--root"):
             root_project_dir = arg